+++ CHANGELOG.md
@@ -518,3 +518,4 @@
 - 2026-04-30: Features mod, summaries
 - 2026-04-30: Err summaries, mod
 - 2026-05-01: Features mod, summaries
+- 2026-05-02: Add mod, summaries types